aggregator market info line moved to new app
